+++ app/forms/transfer/edit_project_form.rb
@@ -8,17 +8,22 @@ class Transfer::EditProjectForm
   attribute :establishment_sharepoint_link
   attribute :outgoing_trust_sharepoint_link
   attribute :incoming_trust_sharepoint_link
+  attribute :outgoing_trust_ukprn
 
   validates :establishment_sharepoint_link, presence: true, sharepoint_url: true
   validates :incoming_trust_sharepoint_link, presence: true, sharepoint_url: true
   validates :outgoing_trust_sharepoint_link, presence: true, sharepoint_url: true
+
+  validates :outgoing_trust_ukprn, presence: true, ukprn: true
+  validates :outgoing_trust_ukprn, trust_exists: true, if: -> { outgoing_trust_ukprn.present? }
 
   def self.new_from_project(project)
     new(
       project: project,
       establishment_sharepoint_link: project.establishment_sharepoint_link,
       incoming_trust_sharepoint_link: project.incoming_trust_sharepoint_link,
-      outgoing_trust_sharepoint_link: project.outgoing_trust_sharepoint_link
+      outgoing_trust_sharepoint_link: project.outgoing_trust_sharepoint_link,
+      outgoing_trust_ukprn: project.outgoing_trust_ukprn
     )
   end
 
@@ -30,7 +35,8 @@ class Transfer::EditProjectForm
     project.assign_attributes(
       establishment_sharepoint_link: establishment_sharepoint_link,
       incoming_trust_sharepoint_link: incoming_trust_sharepoint_link,
-      outgoing_trust_sharepoint_link: outgoing_trust_sharepoint_link
+      outgoing_trust_sharepoint_link: outgoing_trust_sharepoint_link,
+      outgoing_trust_ukprn: outgoing_trust_ukprn
     )
     if valid?
       project.save